Run keycloak with a real, persistant db
Wow, I am such a dumb dumb! I never configured a persistent store for
keycloak, and we just lost all it's configuration when moving :cry:

We've now got keycloak connected to a real postgres db, and I also made
the other tweaks necessary to get it running in so-called "production"
while I was in here.