Add python2 to asdf-nix
THIS WAS AN INSANE AMOUNT OF WORK. I had to use pip2nix to get a
the right set of pip dependencies to install virtualenv, and after I
spent an eternity on that, it turns out that virtualenv doesn't even
work with nix's python2 setup. Thankfully, there's a patch (linked to
below) that address *that* issue.