Cleanup fonts
Wow that was involved, but I'm mostly happy with how this went. Google's
monochrome noto font is nice, but it's a bit of a pain to get right now =(

I also don't understand why the batteries look so tiny with the
UbuntuMono Nerd Font, so I've got this weird hack to make that look
nice.