Remove docker wrapper with conf
I was pretty happy with the old, non-home-manager solution, but `docker
compose` is failing because it's trying to write to the conf dir
(grr...).

    $ docker compose build --pull
    mkdir /nix/store/whjina6hikncjqd4f6qj0ydyihi6jajd-config.json/buildx: read-only file system

There's probably some way of preventing this, but for now I'm just going
to go back to a regular old `~/.docker` directory and have home-manager
manage just a `config.json` symlink inside of it.