dallben + clark: Do not block boot on waiting for nfs
I'm starting to use this slick automount functionality so machines can get
further in their boot even if our nas is down. This made things a little
tricky for dallben though, because that means the machine boots up so
quickly it isn't even fully online by the time kodi boots up, and then
kodi quickly fails to connect to the mysql server running on clark.

I solved this with a `ExecStartPre` to get kodi to wait until we can
actually connect to the mysql server running on clark.