Add support for openvpn3
Unfortunately, networkmanager only supports openvpn2 which can't handle
webauth SAML based authentication, so we've gotta start doing this
ourselves.

I didn't look into how much work it would be to add support for this to
NetworkManager-openvpn itself, or even to write a simple networkmanager
plugin to handle this. Instead, I wrote a simple script for toggling the
vpn, and a polybar script to show the vpn status.