Urg relative flake references are a *pain*
I wish I had never tried to do this, so many silly bugs where I wasn't
running the code I thought I was. The final straw was finding that this
doesn't even work from scratch on a machine that hasn't already added
the relevative path to its store. Giving up on this for now. I'm
subscribed to the relevant upstream issue, maybe this will get better in
the future.