Fix docker networking
Oops, I broke DNS in containers in
https://github.com/jfly/snow/commit/72a439edf8eefc0a91fad8c0353d8661b5c76c52

This was was sort of subtle! Somewhat relevant:
https://superuser.com/questions/1702091/how-should-systemd-resolved-and-docker-interact.

Docker is smart enough to detect that we're using sytemd-resolved. It
then pulls our DNS servers from /run/systemd/resolve/resolv.conf, which
includes dnsmasq, which was previously *only* listening to the loopback
interface, and thereby ignoring all dns queries coming from the docker
network. I added the docker network to the list, and all is well.

See comment below about an alternative solution involing `bind-dynamic`
that might be preferable if this list gets annoying to maintain.