Start implementing more granular snowauth authentication
This was pretty involved, and it's not really great. We're just creating
a traefik middleware per set of users that should have access to a given
set of apps.

I used this as an excuse to port more stuff from yaml to pulumi (and I
ported the installation of cert-manager from bash to yaml).