Upgrade to transmission 4
That was a fun adventure! This reminded me that transmission (bizarrely)
chooses to store the list of ongoing downloads in its config directory.
This feels like state to me rather than config.

Maybe it would be good to switch to a different torrenting tool.
rTorrent + flood looks like a pretty slick combination.