Get rid of eza
It's fine (nearly identical to `ls`, I think the only different I found
was `ls symlink` vs `ls symlink/`, which behaves differently with
regular ls, but the same with eza).

I'm getting rid of this just because it seems to mess with my
completion. I'm sure it's not eza's fault, but this is the easiest way
to fix things.