agenix-rooter: don't assume the agenix activationScript exists
This is analogous to this if statement in agenix:
https://github.com/ryantm/agenix/blob/13ac9ac6d68b9a0896e3d43a082947233189e247/modules/age.nix#L254,
which prevents `system.activationScripts.agenix` from getting defined,
and if it's not defined, then there's nothing to do.

This might not be quite right: don't we need to clean up the secrets
when we remove the last secret? Does agenix have a similar bug?