Revert "Turn on ipv6 relay mode"
This reverts commit 8cda10856ed6d1c9d12035b5c34900905384867c.

This was breaking DNS on our network: the router would send clients the
"upstream" ipv6 dns server, which wasn't our router, which would not
know how to resolve dns entries for `*.lan` addresses.

I tried reverting this via the Luci ui and everything seems to work,
including ipv6 to the outside world (which I said wasn't working with my
original change). So, I'm a bit stumped...