More refactoring
Get rid of the `nix/` folder, we can accomplish all that with the
existing `lib/` folder.

I also renamed `pkgs/` to `packages/`. It's consistent with the name
of the flake output, and "pkgs" just sounds like nixpkgs to me now.

I've also reworked things so we're more consistent about how we handle
`systemArgs`. Next I want to introduce a `lib'` construct to genericize
this "system specific nix lib" pattern.