Port pattern to clan
This turned into quite a mess: AFAICT, clan doesn't have support for
templated secrets (or maybe it does, because I know it's using sops-nix
under the hood, which does have support for templated secrets?). pattern
leveraged this to generate a docker config with credentials to push to
the OCI registry running on k8s.

I took this as an opportunity to remove the OCI registry from k8s. It
makes for a weird circular dependency. Now, you just have to be on the
same network as clark.ec and you can push to it unauthenticated. Not
ideal, but I plan for this to be short-lived, as I intend to tear down
our k8s cluster.